JINS MEME notes
Press for relationship with JINS MEME.  For it to find use it must be able to snaptshot itself in Cain space.  That requires that it sense something to anchor to.  Any GPS bird for which we know its position in Cain space would do to distribute real time value of Cain.  Birds must know their interval from other birds as a matter of course so distribution of current value of Cain down to the hand (face) is a piece of cake.